update according to PR review commentaries
